% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dskewnorm.R
\name{dskewnorm}
\alias{dskewnorm}
\title{Density function of the Skew-Normal Distribution}
\usage{
dskewnorm(x, par1 = 0, par2 = 1, par3 = 0)
}
\arguments{
\item{x}{A vector of quantiles.}

\item{par1}{Location parameter of the skew-normal distribution.}

\item{par2}{Scale parameter of the skew-normal distribution.}

\item{par3}{Slant parameter of the skew-normal distribution.}
}
\description{
Density function of the Skew-Normal Distribution
}
\examples{
dskewnorm(x = c(-0.5,0.9))

}
\references{
Owen, D. B. (1956). Tables for Computing Bivariate Normal Probabilities.
\emph{The Annals of Mathematical Statistics}, \strong{27}(4), 1075–1090.

Azzalini, A., and Capitanio, A. (2014). The Skew-Normal and Related Families.
\emph{Cambridge University Press}.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
