% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimSI.R
\name{estimSI}
\alias{estimSI}
\title{Nonparametric serial interval estimation}
\usage{
estimSI(x, nboot = 2000)
}
\arguments{
\item{x}{A data frame with n rows (number of transmission pairs) and two
columns: the left and right bound, respectively, of the observed serial
interval window.}

\item{nboot}{The bootstrap sample size to construct confidence intervals.}
}
\value{
A list containing the parameter estimates.
}
\description{
This routine uses a nonparametric methodology to estimate the serial interval
distribution based on interval censored serial interval data. It requires as
input a data frame with two columns (the left and right bound of the obseved
serial interval window). The routine outputs point estimates and confidence
intervals for selected features of the serial interval that are often reported.
}
\details{
The \code{estimSI} routine provides point estimates and confidence
intervals for the mean, standard deviation, and 5th, 25th,
50th (median), 75th and 95th quantiles of the serial interval distribution.
}
\examples{
# Add examples here
# estimSI...

}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
