% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerboot.R
\name{kerboot}
\alias{kerboot}
\title{Bootstrap resampling routine}
\usage{
kerboot(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
A data frame representing a bootstrap sample of `x`. The number of
bootstrap samples equals the number of rows of `x`.
}
\description{
Resample with replacement rows of a data frame.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
\keyword{internal}
