% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerdata_check.R
\name{kerdata_check}
\alias{kerdata_check}
\title{Internal check of data frame provided by user}
\usage{
kerdata_check(x)
}
\arguments{
\item{x}{A data frame.}
}
\value{
A list with "pass" (if `x` is in correct format) or "fail" if
(`x` is not in correct format) and a message.
}
\description{
Checks if the data frame provided by user is in correct format.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
\keyword{internal}
