% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerdomain_check.R
\name{kerdomain_check}
\alias{kerdomain_check}
\title{Check if the sample space of parametric family is compatible with data}
\usage{
kerdomain_check(x, family)
}
\arguments{
\item{x}{A data frame with either two columns named \code{xl} and \code{xr},
or four columns named \code{x1l}, \code{x1r}, \code{x2l}, \code{x2r}.}

\item{family}{A character string specifying the name of the parametric
family.}
}
\description{
Check if the sample space of parametric family is compatible with data
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
\keyword{internal}
