% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerfamilies.R
\name{kerfamilies}
\alias{kerfamilies}
\title{Description of available parametric families}
\usage{
kerfamilies()
}
\value{
A list summarizing available parametric families, the number of
parameters involved, the parameter names and the support.
}
\description{
Provides a summary of available parametric families.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
\keyword{internal}
