% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerfamily_check.R
\name{kerfamily_check}
\alias{kerfamily_check}
\title{Internal check of parametric family name provided by user}
\usage{
kerfamily_check(x)
}
\arguments{
\item{x}{A string representing a parametric family name.}
}
\value{
A list with "pass" (if `x` is an available family) or "fail" if
(`x` is not an available family) and a message.
}
\description{
Checks if the parametric family name specified by user is available.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
\keyword{internal}
