% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerstats_boot.R
\name{kerstats_boot}
\alias{kerstats_boot}
\title{Computes summary statistics}
\usage{
kerstats_boot(x, pe, boot)
}
\arguments{
\item{x}{A list of entries for which the statistics are required.}

\item{pe}{A vector of point estimates.}

\item{boot}{Bootstraped statistics.}
}
\description{
Computes summary statistics
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
\keyword{internal}
