% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerstats_norm.R
\name{kerstats_norm}
\alias{kerstats_norm}
\title{Computes summary statistics}
\usage{
kerstats_norm(x, pe, se)
}
\arguments{
\item{x}{A list of entries for which the statistics are required.}

\item{pe}{A vector of point estimates.}

\item{se}{Estimated standard error.}
}
\description{
Computes summary statistics
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
\keyword{internal}
