% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerTOw.R
\name{kerTOw}
\alias{kerTOw}
\title{Evaluation of Owen's \eqn{T(h,a)} function}
\usage{
kerTOw(h, a, kmax = 50)
}
\arguments{
\item{h}{A real number.}

\item{a}{A real number.}

\item{kmax}{An integer to truncate the infinite series expression of \eqn{T(h,a)}.}
}
\value{
A real number.
}
\description{
Evaluates the function \eqn{T(h,a)} described in Owen (1956).
}
\references{
Owen, D. B. (1956). Tables for Computing Bivariate Normal Probabilities.
\emph{The Annals of Mathematical Statistics}, \strong{27}(4), 1075–1090.

Azzalini, A., and Capitanio, A. (2014). The Skew-Normal and Related Families.
\emph{Cambridge University Press}.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
\keyword{internal}
