% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logliksingle.R
\name{logliksingle}
\alias{logliksingle}
\title{Log likelihood functions and model features for single interval-censored data}
\usage{
logliksingle(x, family)
}
\arguments{
\item{x}{A data frame with n rows and two columns indicating the left and
right bound, respectively, of the interval-censored observations. \code{NA}
values are not allowed.}

\item{family}{The name of the parametric family used in modeling the data.
Must be one of: "gamma", "lognormal", "weibull" or "gaussian".}
}
\value{
A list containing the log likelihood function and other model
 features.
}
\description{
Provides the log likelihood function, features, and parameter space for a
given parametric family.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr} and
 Dongxuan Chen.
}
\keyword{internal}
