% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midmom.R
\name{midmom}
\alias{midmom}
\title{Compute initial values for maximum likelihood estimation via moment matching}
\usage{
midmom(x, family = "gamma")
}
\arguments{
\item{x}{A data frame with n rows and two columns indicating the left and
right bound, respectively, of the interval-censored observations. \code{NA}
values are not allowed.}

\item{family}{The name of the parametric family used in modeling the data.
Must be one of: "gamma", "lognormal", "weibull" or "gaussian".}
}
\value{
A vector of parameter values.
}
\description{
Compute initial values for maximum likelihood estimation via moment matching
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
\keyword{internal}
