% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparfit.R
\name{nonparfit}
\alias{nonparfit}
\title{Nonparametric estimation for single interval-censored data based on uniform mixtures}
\usage{
nonparfit(x, Bboot = 1000, p = NULL, pgbar = TRUE)
}
\arguments{
\item{x}{A data frame with n rows and two columns indicating the left and
right bound, respectively, of the interval-censored observations. \code{NA}
values are not allowed.}

\item{Bboot}{Number of bootstrap samples. Default is 1000.}

\item{p}{A vector of probabilities. An associated quantile estimate will be
computed for each entry of p based on the fitted epidemiological delay
distribution. By default p = c(0.01, 0.05, 0.25, 0.50, 0.75, 0.95, 0.99).}

\item{pgbar}{Should a progress bar be displayed in console? Default is TRUE.}
}
\value{
A list containing detailed information on the fitted nonparametric
model. The \code{summary()} function can be used to see further details.
}
\description{
This routine uses a nonparametric methodology based on uniform mixtures
(Gressani and Hens, 2025) to estimate different distributional features from
interval-censored data. It is particularly useful in an epidemiological delay
modeling context to extract information from coarse delay data without
imposing any distributional assumption on the underlying data generating
mechanism. The routine requires as main input a data frame \code{x} with two
columns (named \code{xl} and \code{xr}) indicating the left and right bound,
respectively, of the interval-censored observations. The naming convention of
data frame columns is strict and different namings are not allowed.
Note that the left bound should be strictly smaller than the right
bound, i.e. \code{xl < xr} must be satisfied for all rows in \code{x}.
The routine computes point estimates and confidence intervals
(via the nonparametric bootstrap) for different features of the underlying
random variable being modeled. The following features are considered: mean,
variance, standard deviation, and 1\%, 5\%, 25\%, 50\%, 75\%, 95\% and 99\%
percentiles.
}
\examples{
# Add examples here

}
\references{
Gressani, O. and Hens, N. (2025). Nonparametric serial interval
estimation with uniform mixtures. \emph{Plos Computational Biology},
 \strong{21}(8): e1013338.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
