% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_plot_nonparfit.R
\name{plot.nonparfit}
\alias{plot.nonparfit}
\title{Plot the cumulative distribution function of a fitted nonparametric model}
\usage{
\method{plot}{nonparfit}(x, xlim = NULL, grid = 500L, legend = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{nonparfit}.}

\item{xlim}{A numeric vector of size 2 giving the x-axis range. Default left
bound is given by the minimum of the observed left bound and default right
bound is given by the maximum of the observed right bound.}

\item{grid}{The number of grid points on which to evaluate the fitted
distribution.}

\item{legend}{Add a legend?}

\item{...}{Further graphical parameters to be passed to the plot function.}
}
\description{
Can be used to plot the cumulative distribution function of a
nonparametric model fitted with the \code{nonparfit} routine.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
