% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_plot_parfitml.R
\name{plot.parfitml}
\alias{plot.parfitml}
\title{Plot the density or cumulative distribution function of a fitted
parametric model}
\usage{
\method{plot}{parfitml}(x, xlim = NULL, grid = 500L, target = c("pdf", "cdf"), legend = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{parfitml}.}

\item{xlim}{A numeric vector of size 2 giving the x-axis range. Default left
bound is given by the minimum of the observed left bound and default right
bound is given by the maximum of the observed right bound.}

\item{grid}{The number of grid points on which to evaluate the fitted
distribution.}

\item{target}{The fitted target function to be plotted, `pdf` is the default
and gives the probability density function. Specifying `cdf` plots the fitted
cumulative distribution function.}

\item{legend}{Add a legend?}

\item{...}{Further graphical parameters to be passed to the plot function.}
}
\description{
Can be used to plot the probability density function
or the cumulative distribution function of a parametric model fitted with
the \code{parfitml} routine.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
