% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSI.R
\name{simSI}
\alias{simSI}
\title{Simulation of artificial serial interval data}
\usage{
simSI(muS = 3, sdS = 2, maxcoarse = 3, probcoarse = c(0.8,0.15,0.05), n = 10)
}
\arguments{
\item{muS}{The mean of the target serial interval distribution.}

\item{sdS}{The standard deviation of the serial interval distribution.}

\item{maxcoarse}{The largest degree of coarseness allowed (in days).}

\item{probcoarse}{The (discrete) censoring probability distribution (must sum to one).}

\item{n}{The sample size (number of transmission pairs).}
}
\value{
A list with the following components:
\itemize{
 \item s: The true serial interval value.
 \item sl: Left bound of the serial interval window.
 \item sr: Right bound of the serial interval window.
 \item sw: Width of the serial interval window.
}
}
\description{
This routine simulates artificial serial interval data with serial
interval windows having a width of at least two days. It assumes that the
target serial interval (SI) distribution is Gaussian. The width of the
serial interval windows can be controlled by specifying a discrete censoring
distribution.
}
\details{
The \code{simSI} routine outputs a data frame containing the
 generated serial interval values (which are in practice unobserved), the left
 and right bound of the serial windows (observed) and the width of the
 serial interval windows.
}
\examples{
# Simulate serial interval windows for n=15 transmission pairs
sigen <- simSI(n = 15)

}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
