% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_summary_nonparfit.R
\name{summary.nonparfit}
\alias{summary.nonparfit}
\title{Summarize nonparametric model fit}
\usage{
\method{summary}{nonparfit}(object, ndigits = 3, ...)
}
\arguments{
\item{object}{An object of class \code{nonparfit}.}

\item{ndigits}{Number of digits to print in output.}

\item{...}{Further arguments to be passed.}
}
\value{
A summary output.
}
\description{
This routine can be used to summarize a nonparametric model fit.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
