% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_summary_parfitml.R
\name{summary.parfitml}
\alias{summary.parfitml}
\title{Summarize parametric model fit with maximum likelihood}
\usage{
\method{summary}{parfitml}(object, ndigits = 3, type = "full", ...)
}
\arguments{
\item{object}{An object of class \code{parfitml}.}

\item{ndigits}{Number of digits to print in output.}

\item{type}{Type of summary output. Either \code{"full"} or \code{"compact"}.}

\item{...}{Further arguments to be passed.}
}
\value{
A summary output.
}
\description{
This routine can be used to summarize a parametric model fit
obtained with maximum likelihood.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
