% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerfeats.R
\name{kerfeats}
\alias{kerfeats}
\title{Computes features for a given parametric family}
\usage{
kerfeats(family, par)
}
\arguments{
\item{family}{A characther string specifying the parametric family.}

\item{par}{A vector of parameter values.}
}
\value{
A data frame of features.
}
\description{
Computes the mean, variance, standard deviation and different quantiles
given a parametric family and parameter values.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
\keyword{internal}
