% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerlikelihood.R
\name{kerlikelihood}
\alias{kerlikelihood}
\title{Likelihood function for single and doubly interval-censored data}
\usage{
kerlikelihood(x, family)
}
\arguments{
\item{x}{A data frame with either two columns named \code{xl} and \code{xr},
or four columns named \code{x1l}, \code{x1r}, \code{x2l}, \code{x2r}. See
description for constraints imposed on the columns.}

\item{family}{A character string specifying the name of the parametric
family.}
}
\value{
A list containing information on the chosen parametric family,
the log-likelihood function, a function that transforms back the parameters
in their original scale, and the censoring type. It also
returns the Jacobian of the inverse of the function that transforms a
bounded domain to an unbounded domain.
}
\description{
The likelihood function plays a key role in maximum likelihood estimation and
Bayesian inference. This routine admits as data input a data frame \code{x}
with either two columns (named \code{xl} and \code{xr}) representing the left
and right bound, respectively, of the delay variable, or four columns
(named \code{x1l}, \code{x1r}, \code{x2l}, \code{x2r}) representing the left
and right bound, respectively, of the primary and secondary events of the
delay variable. The naming convention of the columns is strict and different
namings are not allowed. When data frame \code{x} has two columns, a single
interval-censored likelihood function is used. Note that the left bound
should be strictly smaller than the right bound, i.e. \code{xl < xr} must
be satisfied for all rows in \code{x}. When data frame \code{x} has four
columns, a doubly interval-censored likelihood function is used. In that case
\code{x1l < x1r} and \code{x2l < x2r} must hold for all rows in \code{x}.
Moreover, \code{NA} values are not allowed in data frame \code{x}.
}
\details{
Returns the log-likelihood function of parameters that have been transformed
to live in an unbounded parameter space. This makes the
maximization/evaluation of the likelihood function numerically more stable.
The doubly interval-censored likelihood function requires to evaluate an
inner integral. By default, this is implemented via numerical integration,
where the integral is evaluated with the \code{integrate} routine. This
routine also returns a function (logliki) that evaluates the log-likelihood
pointwise at each datum.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
\keyword{internal}
