% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kerstats.R
\name{kerstats}
\alias{kerstats}
\title{Computes summary statistics based on point estimates and bootstrap samples}
\usage{
kerstats(slist, pestim, boot)
}
\arguments{
\item{slist}{A list of bootstrap statistics.}

\item{pestim}{A vector of point estimates.}

\item{boot}{Bootstraped statistics.}
}
\description{
Computes summary statistics based on point estimates and bootstrap samples
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
\keyword{internal}
