% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparfit.R
\name{nonparfit}
\alias{nonparfit}
\title{Nonparametric estimation for interval-censored data based on uniform mixtures}
\usage{
nonparfit(x, boot = FALSE, Bboot = 200, pgbar = FALSE)
}
\arguments{
\item{x}{A data frame with n rows and two columns indicating the left and
right bound, respectively, of the interval-censored observations. \code{NA}
values are not allowed.}

\item{boot}{Should the nonparametric bootstrap be implemented to obtain
standard errors and 90\% and 95\% confidence intervals for different
features of the distribution? Default is \code{FALSE}.}

\item{Bboot}{The number of bootstrap replications. Default is 200.}

\item{pgbar}{Should a progress bar be displayed while the bootstrap is
ongoing? Default is \code{FALSE}.}
}
\value{
A list containing different feature estimates of the underlying
 distribution.
}
\description{
This routine uses a nonparametric methodology based on uniform mixtures
(Gressani and Hens, 2025) to estimate different distributional features from
interval-censored data. It is particularly useful in an epidemiological delay
modeling context to extract information from coarse delay data without
imposing any distributional assumption on the underlying data generating
mechanism. The routine requires as main input a data frame with two columns
indicating the left and right bound, respectively, of the interval-censored
observations. It provides point estimates and confidence intervals
(via the nonparametric bootstrap) for different features of the underlying
random variable being modeled. The following features are considered: mean,
variance, standard deviation, and 1\%, 5\%, 25\%, 50\%, 75\%, 95\% and 99\%
percentiles.
}
\examples{
# Add examples here

}
\references{
Gressani, O. and Hens, N. (2025). Nonparametric serial interval
estimation with uniform mixtures. \emph{Plos Computational Biology},
 \strong{21}(8): e1013338.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
