% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parfitml.R
\name{parfitml}
\alias{parfitml}
\title{Parametric estimation with maximum likelihood for single or doubly
interval-censored data}
\usage{
parfitml(x, family, ci = c("npboot", "pboot", "sbnorm"), ...)
}
\arguments{
\item{x}{A data frame with either two columns named \code{xl} and \code{xr},
or four columns named \code{x1l}, \code{x1r}, \code{x2l}, \code{x2r}. See
description for constraints imposed on the columns.}

\item{family}{A character string specifying the name of the parametric
family. Can be one of the following: \code{"gaussian"}, \code{"gamma"},
\code{"lognormal"}, \code{"weibull"}, or \code{"skewnorm"}.}

\item{ci}{The engine for computing confidence intervals. Default is
\code{"npboot"} for nonparametric bootstrap. Other options are \code{"pboot"}
for the parametric bootstrap (currently under construction) or \code{"sbnorm"}
which relies on a simulation-based approach to compute confidence intervals
using asymptotic normality of the MLE estimator following Mandel (2013).}

\item{...}{Further arguments. Specifying e.g. \code{Bboot = 1000} permits
to fix the bootstrap sample size to 1000 (the default is 100 here). Likewise
specifying \code{ns = 1000} permits to fix the number of simulated samples
from the Gaussian distribution of the MLE in the Mandel (2013) method (the
default is 100 here). Also, specifying \code{pgbar = TRUE} shows a progress
bar for the bootstrap.}
}
\value{
A list containing detailed information on the fitted parametric
model. The \code{summary()} function can be used to see further details.
}
\description{
Fit parametric models to single or doubly
interval-censored data with maximum likelihood. Data input must be a
data frame \code{x} with either two columns (named \code{xl} and \code{xr})
representing the left and right bound, respectively, of the delay variable,
or four columns (named \code{x1l}, \code{x1r}, \code{x2l}, \code{x2r})
representing the left and right bound, respectively, of the primary and
secondary events of the delay variable. The naming convention of the data
columns is strict and different namings are not allowed. When data frame
\code{x} has two columns, a single interval-censored likelihood function is
used. Note that the left bound should be strictly smaller than the right
bound, i.e. \code{xl < xr} must be satisfied for all rows in \code{x}.
When data frame \code{x} has four columns, a doubly interval-censored
likelihood function is used. In that case \code{x1l < x1r} and
\code{x2l < x2r} must hold for all rows in \code{x}. \code{NA} values are
not allowed in data frame \code{x}.
}
\details{
This routine computes maximum likelihood estimates of model
parameters as well as different features of the fitted delay variable
distribution (mean, variance, standard deviation and selected quantiles). The
nonparametric bootstrap is used to compute standard errors and confidence
intervals for the model parameters and different features. Maximum likelihood
estimates are computed with the \code{optim} function using the 
Nelder and Mead (1965) method. Initial parameter values are computed via a
moment matching approach.
}
\references{
Nelder, J. A. and Mead, R. (1965). A simplex algorithm for
function minimization. \emph{Computer Journal}, \strong{7}, 308–313.

Mandel, M. (2013). Simulation-based confidence intervals 
for functions with complicated derivatives. \emph{The American Statistician},
\strong{67}(2), 76-81.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr} (original
code writing and implementation) and
Dongxuan Chen (code editing and testing).
}
