% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parfitml.R
\name{parfitml}
\alias{parfitml}
\title{Parametric estimation with maximum likelihood for interval-censored data}
\usage{
parfitml(x, family = "gamma", boot = FALSE, Bboot = 200, pgbar = FALSE)
}
\arguments{
\item{x}{A data frame with n rows and two columns indicating the left and
right bound, respectively, of the interval-censored observations. \code{NA}
values are not allowed.}

\item{family}{The name of the parametric family used in modeling the data.
Must be one of: "gamma", "lognormal", "weibull" or "gaussian".}

\item{boot}{Should the bootstrap be implemented to obtain standard errors
and 90\% and 95\% confidence intervals for different features of the
distribution? Default is \code{FALSE}.}

\item{Bboot}{The number of bootstrap replications. Default is 200.}

\item{pgbar}{Should a progress bar be displayed while the bootstrap is
ongoing? Default is \code{FALSE}.}
}
\value{
A list containing parameter estimates, feature estimates, AIC, BIC,
and convergence checks.
}
\description{
This routine fits commonly used parametric models to interval-censored data.
}
\details{
The \code{parfitml} routine provides point estimates and confidence
intervals (if bootstrap is called) for different features of the chosen
parametric distribution. The following features are considered: mean,
variance, standard deviation, and 1\%, 5\%, 25\%, 50\%, 75\%, 95\% and 99\%
percentiles. Maximum likelihood estimates of model parameters are computed
with the \code{optim} function with the Nelder and Mead (1965) method.
Initial parameter values are obtained via a moment matching approach, where
the theoretical mean and variance of the chosen parametric model are matched
with the empirical mean and variance of the midpoint imputed values computed
from the interval-censored observations.
}
\examples{
# Add examples here

}
\references{
Nelder, J. A. and Mead, R. (1965). A simplex algorithm for
function minimization. \emph{Computer Journal}, \strong{7}, 308–313.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr} and
 Dongxuan Chen.
}
