% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qskewnorm.R
\name{qskewnorm}
\alias{qskewnorm}
\title{Quantile function of the Skew-Normal Distribution}
\usage{
qskewnorm(p, par1 = 0, par2 = 1, par3 = 0)
}
\arguments{
\item{p}{A vector of probabilities.}

\item{par1}{Location parameter of the skew-normal distribution.}

\item{par2}{Scale parameter of the skew-normal distribution.}

\item{par3}{Slant parameter of the skew-normal distribution.}
}
\description{
Quantile function of the Skew-Normal Distribution
}
\details{
The \code{qskewnorm} routine use Owen's (1956)
\eqn{T(h,a)} function and a \code{uniroot} procedure for quantile
computation.
}
\examples{
qskewnorm(p = c(0.1,0.9))

}
\references{
Owen, D. B. (1956). Tables for Computing Bivariate Normal Probabilities.
\emph{The Annals of Mathematical Statistics}, \strong{27}(4), 1075–1090.

Azzalini, A., and Capitanio, A. (2014). The Skew-Normal and Related Families.
\emph{Cambridge University Press}.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}
}
