% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Simulation from a multivariate Gaussian distribution}
\usage{
rmvnorm(n = 1, mean, sigma)
}
\arguments{
\item{n}{A positive integer indicating the number of draws.}

\item{mean}{The mean vector of the multivariate Gaussian.}

\item{sigma}{The covariance matrix of the multivariate Gaussian distribution.}
}
\value{
A list containing the simulated draws and the method used for 
covariance matrix decomposition.
}
\description{
Simulates random draws from a multivariate normal distribution with mean 
vector \code{mean} and covariance matrix \code{sigma}. The number of draws
is specfied by \code{n} (a positive integer). The routine first tries to 
compute the Cholesky factorization of the covariance matrix by assuming that
the latter is a symmetric positive-definite matrix. If this approach fails,
the routine relies on an eigendecomposition of the covariance matrix as a 
fallback option.
}
\references{
Ripley, B. D. (2009). Stochastic simulation. \emph{John Wiley & Sons}.

Rubinstein, R. Y. and Kroese, D. P. (2016). Simulation and
the Monte Carlo method. \emph{John Wiley & Sons}.
}
\author{
Oswaldo Gressani \email{oswaldo_gressani@hotmail.fr}.
}
